prices/fees changed
